LayoutTests/http/tests/websocket/tests/hixie76/url-with-query-for-no-query.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure handshake with URL with query components fails against server that doesn't support query component.");

window.jsTestIsAsync = true;

var url = "ws://127.0.0.1:8880/websocket/tests/hixie76/no-query?";
var handshake_success = false;
var ws_location;

function endTest()
{
    clearTimeout(timeoutID);
    shouldBeFalse("handshake_success");
    shouldBeUndefined("ws_location");
    finishJSTest();
}

debug("url=" + url);
var ws = new WebSocket(url);
ws.onopen = function () {
    debug("WebSocket is open");
    handshake_success = true;
};
ws.onmessge = function (evt) {
    ws_location = evt.data;
    debug("received:" + ws_location);
    ws.close();
};
ws.onclose = function () {
    debug("WebSocket is closed");
    endTest();
};
var timeoutID = setTimeout("endTest()", 2000);

var successfullyParsed = true;
</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
